% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aapl.R
\docType{data}
\name{aapl}
\alias{aapl}
\title{AAPL market data}
\format{
An object of class \code{"data.frame"}
\describe{
 \item{symbol}{Stock symbol}
 \item{datetime}{Datetime of the bar}
 \item{open}{Open price}
 \item{high}{High price}
 \item{low}{Low price}
 \item{close}{Close price}
 \item{volume}{Volume}
}
}
\usage{
data(aapl)
}
\description{
Market data for AAPL with hour frequency for period from 2004-01-01 to 2021-05-01.
Data was generated through FMP cloud API.
}
\examples{

data(aapl)
head(aapl)

}
\references{
Data was collected using FMP cloud API
}
\keyword{datasets}
