\name{CCM.pvals}
\alias{CCM.pvals}
\docType{data}
\title{
  Gene p-value list derived from knock-out experiments of the three CCM genes
}
\description{
  Gene p-value list derived from knock-out experiments of the three CCM genes CCM1, CCM2 and CCM3.
  Contains p-values obtained from a limma differential expression analysis of knock-out samples versus control samples (each done in triplicate).
}
\usage{data("CCM.pvals")}
\format{
  Named list (CCM1,2,3) of named numericals (names = gene ids, values = limma p-values) 
}
\keyword{datasets}
