% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factor_file.R
\name{create_factor_file}
\alias{create_factor_file}
\title{Get factor files}
\usage{
create_factor_file(ticker, ohlcv, api_key, save_dir)
}
\arguments{
\item{ticker}{Stock ticker}

\item{ohlcv}{Data table with OHLCV columns}

\item{api_key}{FMP cloud API key}

\item{save_dir}{Dir path to save factor files}
}
\value{
Stock IPO date
}
\description{
Get factor files for Lean engine.
}
