% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_market_equities_minutes.R
\name{dhms}
\alias{dhms}
\title{Convert seconds to hour:minute:seconds}
\usage{
dhms(t)
}
\arguments{
\item{t}{Seconds from midnight.}
}
\value{
Hour, minute, second format of seconds.
}
\description{
Convert seconds from midnight to hour:minute:seconds time format.
}
