\name{g2}
\alias{g2}
\docType{data}
\title{
  igraph graph object used in examples for function run.lean.fromdata
}
\description{
 igraph graph object used in examples for function run.lean.fromdata.
 Obtained by parsing the STRING v.91 murine interaction network and restricting 
 it to proteins mappable to genes contained on the Affymetrix MouseGene v1.0 ST chip.
}
\usage{data("g2")}
\format{
  The format is:
IGRAPH UNW- 7342 63617 -- 
}
\source{
 STRING v9.1 Mouse filtered for confidence scores >= 0.9
}
\keyword{datasets}
