\name{g_red}
\alias{g_red}
\docType{data}
\title{
  igraph graph object used in unit tests
}
\description{
 igraph graph object used in unit tests.
 Obtained by restricting the graph <g2> to the graph induced by randomly selecting 1500 genes.
}
\usage{data("g_red")}
\format{
  The format is:
IGRAPH UNW- 1500 2818 -- 
}
\source{
 STRING v9.1 Mouse filtered for confidence scores >= 0.9; radnomly reduced to 1500 genes and all interactions between them
}
\keyword{datasets}
