\name{gene.list.scores}
\alias{gene.list.scores}
\docType{data}
\title{
  Gene p-value list used in examples for function run.lean.fromdata
}
\description{
  Gene p-value list used in examples for function run.lean.fromdata
  Contains p-values obtained from a limma differential expression analysis
}
\usage{data("gene.list.scores")}
\format{
  The format is:
 Named num [1:7342] 0.772 0.813 0.979 0.841 0.607 ...
 - attr(*, "names")= chr [1:7342] "10090.ENSMUSP00000000001" "10090.ENSMUSP00000010205" "10090.ENSMUSP00000053818" "10090.ENSMUSP00000000153" ...
}
\keyword{datasets}
