% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ipo_date.R
\name{get_ipo_date}
\alias{get_ipo_date}
\title{Get IPO date}
\usage{
get_ipo_date(ticker, api_key)
}
\arguments{
\item{ticker}{Stock ticker}

\item{api_key}{Fmp cloud API key.}
}
\value{
Stock IPO date
}
\description{
Get IPO date from fmp cloud.
}
