% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_market_equities_minutes.R
\name{get_market_equities_minutes}
\alias{get_market_equities_minutes}
\title{Import local market data with minute frequency}
\usage{
get_market_equities_minutes(path, tickers = NA)
}
\arguments{
\item{path}{Path to local market data.}

\item{tickers}{Stocks to import.}
}
\value{
Market data organized as panel data.
}
\description{
Import market data with lean structure from dictionary with minute frequency.
}
