% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stock_split_factor.R
\name{get_stock_split_factor}
\alias{get_stock_split_factor}
\title{Get split history}
\usage{
get_stock_split_factor(ticker)
}
\arguments{
\item{ticker}{Stock ticker}
}
\value{
Stock splits
}
\description{
Get split data from the web and calculate split factors.
}
