% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ticker_changes.R
\name{get_ticker_changes}
\alias{get_ticker_changes}
\title{Get ticker changes}
\usage{
get_ticker_changes(ticker)
}
\arguments{
\item{ticker}{Stock ticker}
}
\value{
Stock changes
}
\description{
Get ticker changes from the web and link every change to ticker.
}
