% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_lean.R
\name{import_lean}
\alias{import_lean}
\title{Import local market data}
\usage{
import_lean(path, tickers = NA)
}
\arguments{
\item{path}{Path to local market data.}

\item{tickers}{Stocks to import.}
}
\value{
Market data organized as panel data
}
\description{
Import market data with lean structure from dictionary.
}
