\name{pvals_red}
\alias{pvals_red}
\docType{data}
\title{
  Gene p-value list used in unit tests
}
\description{
  Gene p-value list used in unit tests
  Contains p-values obtained on the CCM2 data for a radnom subselection of 1500 genes. 
  To be used in conjunction with the network contained in <g_red>.
}
\usage{data("pvals_red")}
\format{
  The format is:
 Named num [1:1500] 0.5091 0.4833 0.0454 0.0814 0.0324 ...
 - attr(*, "names")= chr [1:1500] "10090.ENSMUSP00000079341" "10090.ENSMUSP00000106951" "10090.ENSMUSP00000045284" "10090.ENSMUSP00000077744" ...
}
\keyword{datasets}
